requirements.txt
pillow
pytesseract
playwright
python-dotenv
tenacity